pyproject.toml
[tool.poetry]
name = "pomodoroz"
version = "0.1.0"
description = ""
authors = ["lucaslima18 <lucas.ala1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
questionary = "^2.0.1"
pyfiglet = "^1.0.2"
dbus-python = "^1.3.2"
plyer = "^2.1.0"
pygame = "^2.5.2"
pydantic = "^2.7.1"
alembic = "^1.13.1"
sqlalchemy = "^2.0.31"
sqlmodel = "^0.0.19"
python-dotenv = "^1.0.1"
coloredlogs = "^15.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
